+++ semester3/hw3/3.2/src/main/java/com/rprtr258/Tank.java
@@ -1,4 +1,4 @@
-package com.rprtr258.client;
+package com.rprtr258;
 
 import javafx.geometry.*;
 import javafx.scene.canvas.GraphicsContext;
